correzione crud

+++ src/main/java/it/contrader/model/Player.java
@@ -24,7 +24,7 @@ public class Player {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+	private long id;
 	
 	
 	private int idPlayer;
